+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/headless/lrop_v4_no_ui5_version/webapp/test/integration/pages/JourneyRunner.js
@@ -1,17 +1,17 @@
 sap.ui.define([
     "sap/fe/test/JourneyRunner",
-	"testnamepsace/lropv4noui5version/test/integration/pages/TravelList",
-	"testnamepsace/lropv4noui5version/test/integration/pages/TravelObjectPage",
-	"testnamepsace/lropv4noui5version/test/integration/pages/BookingObjectPage"
-], function (JourneyRunner, TravelList, TravelObjectPage, BookingObjectPage) {
+	"testnamepsace/lropv4noui5version/test/integration/pages/TravelList.gen",
+	"testnamepsace/lropv4noui5version/test/integration/pages/TravelObjectPage.gen",
+	"testnamepsace/lropv4noui5version/test/integration/pages/BookingObjectPage.gen"
+], function (JourneyRunner, TravelListGenerated, TravelObjectPageGenerated, BookingObjectPageGenerated) {
     'use strict';
 
     var runner = new JourneyRunner({
         launchUrl: sap.ui.require.toUrl('testnamepsace/lropv4noui5version') + '/test/flp.html#app-preview',
         pages: {
-			onTheTravelList: TravelList,
-			onTheTravelObjectPage: TravelObjectPage,
-			onTheBookingObjectPage: BookingObjectPage
+			onTheTravelListGenerated: TravelListGenerated,
+			onTheTravelObjectPageGenerated: TravelObjectPageGenerated,
+			onTheBookingObjectPageGenerated: BookingObjectPageGenerated
         },
         async: true
     });