+++ packages/abap-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @sap-ux/abap-deploy-config-sub-generator
 
+## 1.0.23
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/abap-deploy-config-writer 1.0.9 → 1.0.10
+- @sap-ux/axios-extension 2.0.3 → 2.0.4
+- @sap-ux/feature-toggle 1.0.2 → 1.0.3
+- @sap-ux/adp-tooling 1.0.20 → 1.0.21
+- @sap-ux/btp-utils 2.0.2 → 2.0.3
+- @sap-ux/abap-deploy-config-inquirer 2.0.15 → 2.0.16
+- @sap-ux/deploy-config-generator-shared 1.0.12 → 1.0.13
+- @sap-ux/fiori-generator-shared 1.0.12 → 1.0.13
+- @sap-ux/ui5-config 1.0.3 → 1.0.3
+
+## 1.0.22
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.2 → 2.1.3
+- @sap-ux/abap-deploy-config-writer 1.0.8 → 1.0.9
+- @sap-ux/adp-tooling 1.0.19 → 1.0.20
+- @sap-ux/axios-extension 2.0.3 → 2.0.3
+- @sap-ux/fiori-generator-shared 1.0.11 → 1.0.12
+- @sap-ux/abap-deploy-config-inquirer 2.0.14 → 2.0.15
+- @sap-ux/deploy-config-generator-shared 1.0.11 → 1.0.12
+
+## 1.0.21
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.18 → 1.0.19
+
+## 1.0.20
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.17 → 1.0.18
+
+## 1.0.19
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/abap-deploy-config-inquirer 2.0.13 → 2.0.14
+
 ## 1.0.18
 
 ### Patch Changes
@@ -12,6 +66,8 @@
 
 ## 1.0.17
 
+_Released: 2026-06-16T21:19:36Z_
+
 ### Patch Changes
 
 #### Workspace Updates
@@ -20,6 +76,8 @@
 
 ## 1.0.16
 
+_Released: 2026-06-14T10:40:09Z_
+
 ### Patch Changes
 
 - Updated dependencies [adae40d]
@@ -27,12 +85,16 @@
 
 ## 1.0.15
 
+_Released: 2026-06-12T10:49:08Z_
+
 ### Patch Changes
 
 - Updated dependencies [6e4a71a]
     - @sap-ux/adp-tooling@1.0.15
 
 ## 1.0.14
+
+_Released: 2026-06-12T06:53:23Z_
 
 ### Patch Changes
 
@@ -47,12 +109,16 @@
 
 ## 1.0.13
 
+_Released: 2026-06-11T13:37:16Z_
+
 ### Patch Changes
 
 - Updated dependencies [e66a8a9]
     - @sap-ux/adp-tooling@1.0.13
 
 ## 1.0.12
+
+_Released: 2026-06-11T10:54:17Z_
 
 ### Patch Changes
 
@@ -62,6 +128,8 @@
 - @sap-ux/adp-tooling@1.0.12
 
 ## 1.0.11
+
+_Released: 2026-06-10T09:57:42Z_
 
 ### Patch Changes
 
@@ -77,6 +145,8 @@
 
 ## 1.0.10
 
+_Released: 2026-06-09T14:35:01Z_
+
 ### Patch Changes
 
 - Updated dependencies [0fa8305]
@@ -90,6 +160,8 @@
 
 ## 1.0.9
 
+_Released: 2026-06-09T13:18:16Z_
+
 ### Patch Changes
 
 - Updated dependencies [a328e14]
@@ -99,6 +171,8 @@
     - @sap-ux/adp-tooling@1.0.9
 
 ## 1.0.8
+
+_Released: 2026-06-04T13:54:21Z_
 
 ### Patch Changes
 
@@ -113,12 +187,16 @@
 
 ## 1.0.7
 
+_Released: 2026-06-04T12:10:05Z_
+
 ### Patch Changes
 
 - Updated dependencies [fb84c0e]
     - @sap-ux/adp-tooling@1.0.7
 
 ## 1.0.6
+
+_Released: 2026-06-04T10:19:37Z_
 
 ### Patch Changes
 
@@ -132,6 +210,8 @@
     - @sap-ux/deploy-config-generator-shared@1.0.5
 
 ## 1.0.5
+
+_Released: 2026-06-03T14:58:37Z_
 
 ### Patch Changes
 
@@ -151,6 +231,8 @@
 
 ## 1.0.4
 
+_Released: 2026-06-03T13:52:44Z_
+
 ### Patch Changes
 
 - Updated dependencies [9580241]
@@ -165,6 +247,8 @@
 
 ## 1.0.3
 
+_Released: 2026-06-02T11:35:17Z_
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-inquirer@2.0.2
@@ -174,12 +258,16 @@
 
 ## 1.0.2
 
+_Released: 2026-06-02T08:56:31Z_
+
 ### Patch Changes
 
 - Updated dependencies [a8e4cf0]
     - @sap-ux/adp-tooling@1.0.2
 
 ## 1.0.1
+
+_Released: 2026-06-01T15:15:26Z_
 
 ### Patch Changes
 
@@ -193,6 +281,8 @@
     - @sap-ux/deploy-config-generator-shared@1.0.1
 
 ## 1.0.0
+
+_Released: 2026-05-30T20:54:07Z_
 
 ### Major Changes
 
@@ -242,6 +332,8 @@
 
 ## 0.4.11
 
+_Released: 2026-05-29T12:50:34Z_
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-inquirer@1.9.7
@@ -250,12 +342,16 @@
 
 ## 0.4.10
 
+_Released: 2026-05-29T06:59:27Z_
+
 ### Patch Changes
 
 - Updated dependencies [9a980a9]
     - @sap-ux/adp-tooling@0.19.10
 
 ## 0.4.9
+
+_Released: 2026-05-27T11:39:21Z_
 
 ### Patch Changes
 
@@ -265,6 +361,8 @@
 - @sap-ux/fiori-generator-shared@0.15.6
 
 ## 0.4.8
+
+_Released: 2026-05-27T10:42:47Z_
 
 ### Patch Changes
 
@@ -276,6 +374,8 @@
 - @sap-ux/abap-deploy-config-writer@0.4.4
 
 ## 0.4.7
+
+_Released: 2026-05-26T16:40:21Z_
 
 ### Patch Changes
 
@@ -291,12 +391,16 @@
 
 ## 0.4.6
 
+_Released: 2026-05-22T13:30:05Z_
+
 ### Patch Changes
 
 - Updated dependencies [758b0d4]
     - @sap-ux/adp-tooling@0.19.6
 
 ## 0.4.5
+
+_Released: 2026-05-21T16:21:11Z_
 
 ### Patch Changes
 
@@ -312,12 +416,16 @@
 
 ## 0.4.4
 
+_Released: 2026-05-21T14:58:44Z_
+
 ### Patch Changes
 
 - Updated dependencies [dab1aa2]
     - @sap-ux/adp-tooling@0.19.4
 
 ## 0.4.3
+
+_Released: 2026-05-19T15:16:46Z_
 
 ### Patch Changes
 
@@ -332,12 +440,16 @@
 
 ## 0.4.2
 
+_Released: 2026-05-18T08:15:14Z_
+
 ### Patch Changes
 
 - Updated dependencies [342c544]
     - @sap-ux/adp-tooling@0.19.2
 
 ## 0.4.1
+
+_Released: 2026-05-15T13:12:06Z_
 
 ### Patch Changes
 
@@ -350,6 +462,8 @@
     - @sap-ux/deploy-config-generator-shared@0.2.1
 
 ## 0.4.0
+
+_Released: 2026-05-15T08:12:20Z_
 
 ### Minor Changes
 
@@ -372,6 +486,8 @@
 
 ## 0.3.37
 
+_Released: 2026-05-14T21:28:41Z_
+
 ### Patch Changes
 
 - @sap-ux/adp-tooling@0.18.134
@@ -380,12 +496,16 @@
 
 ## 0.3.36
 
+_Released: 2026-05-14T14:16:50Z_
+
 ### Patch Changes
 
 - Updated dependencies [8c4185a]
     - @sap-ux/adp-tooling@0.18.133
 
 ## 0.3.35
+
+_Released: 2026-05-14T11:45:51Z_
 
 ### Patch Changes
 
@@ -405,6 +525,8 @@
 
 ## 0.3.34
 
+_Released: 2026-05-13T09:36:59Z_
+
 ### Patch Changes
 
 - Updated dependencies [21abda3]
@@ -418,6 +540,8 @@
 
 ## 0.3.33
 
+_Released: 2026-05-12T18:00:39Z_
+
 ### Patch Changes
 
 - Updated dependencies [9360ea5]
@@ -427,6 +551,8 @@
     - @sap-ux/adp-tooling@0.18.130
 
 ## 0.3.32
+
+_Released: 2026-05-06T23:02:00Z_
 
 ### Patch Changes
 
@@ -445,6 +571,8 @@
 
 ## 0.3.31
 
+_Released: 2026-05-01T15:46:09Z_
+
 ### Patch Changes
 
 - Updated dependencies [b2ffc7e]
@@ -452,12 +580,16 @@
 
 ## 0.3.30
 
+_Released: 2026-04-30T19:47:20Z_
+
 ### Patch Changes
 
 - Updated dependencies [fcaa70c]
     - @sap-ux/adp-tooling@0.18.127
 
 ## 0.3.29
+
+_Released: 2026-04-30T14:23:24Z_
 
 ### Patch Changes
 
@@ -472,11 +604,15 @@
 
 ## 0.3.28
 
+_Released: 2026-04-30T13:10:33Z_
+
 ### Patch Changes
 
 - @sap-ux/adp-tooling@0.18.125
 
 ## 0.3.27
+
+_Released: 2026-04-29T15:24:37Z_
 
 ### Patch Changes
 
@@ -492,6 +628,8 @@
 
 ## 0.3.26
 
+_Released: 2026-04-27T19:47:46Z_
+
 ### Patch Changes
 
 - Updated dependencies [1d60871]
@@ -505,12 +643,16 @@
 
 ## 0.3.25
 
+_Released: 2026-04-27T07:30:24Z_
+
 ### Patch Changes
 
 - Updated dependencies [10847a1]
     - @sap-ux/adp-tooling@0.18.122
 
 ## 0.3.24
+
+_Released: 2026-04-23T12:54:21Z_
 
 ### Patch Changes
 
@@ -525,6 +667,8 @@
 
 ## 0.3.23
 
+_Released: 2026-04-23T06:48:55Z_
+
 ### Patch Changes
 
 - Updated dependencies [237371b]
@@ -537,6 +681,8 @@
 
 ## 0.3.22
 
+_Released: 2026-04-22T12:38:46Z_
+
 ### Patch Changes
 
 - @sap-ux/adp-tooling@0.18.119
@@ -544,6 +690,8 @@
 - @sap-ux/deploy-config-generator-shared@0.1.118
 
 ## 0.3.21
+
+_Released: 2026-04-15T11:53:17Z_
 
 ### Patch Changes
 
@@ -555,12 +703,16 @@
 
 ## 0.3.20
 
+_Released: 2026-04-15T08:11:32Z_
+
 ### Patch Changes
 
 - Updated dependencies [8fb08a2]
     - @sap-ux/adp-tooling@0.18.117
 
 ## 0.3.19
+
+_Released: 2026-04-14T20:26:28Z_
 
 ### Patch Changes
 
@@ -574,6 +726,8 @@
     - @sap-ux/abap-deploy-config-writer@0.3.10
 
 ## 0.3.18
+
+_Released: 2026-04-14T12:35:35Z_
 
 ### Patch Changes
 
@@ -590,6 +744,8 @@
 
 ## 0.3.17
 
+_Released: 2026-04-14T11:39:16Z_
+
 ### Patch Changes
 
 - Updated dependencies [497317c]
@@ -597,12 +753,16 @@
 
 ## 0.3.16
 
+_Released: 2026-04-09T07:00:16Z_
+
 ### Patch Changes
 
 - Updated dependencies [7a8613b]
     - @sap-ux/adp-tooling@0.18.113
 
 ## 0.3.15
+
+_Released: 2026-04-08T13:10:18Z_
 
 ### Patch Changes
 
@@ -621,6 +781,8 @@
 
 ## 0.3.14
 
+_Released: 2026-04-07T11:09:34Z_
+
 ### Patch Changes
 
 - Updated dependencies [1b10e9f]
@@ -628,12 +790,16 @@
 
 ## 0.3.13
 
+_Released: 2026-04-01T14:51:40Z_
+
 ### Patch Changes
 
 - Updated dependencies [6b74074]
     - @sap-ux/adp-tooling@0.18.110
 
 ## 0.3.12
+
+_Released: 2026-04-01T11:49:37Z_
 
 ### Patch Changes
 
@@ -648,12 +814,16 @@
 
 ## 0.3.11
 
+_Released: 2026-04-01T06:34:51Z_
+
 ### Patch Changes
 
 - Updated dependencies [68b5523]
     - @sap-ux/adp-tooling@0.18.108
 
 ## 0.3.10
+
+_Released: 2026-03-30T22:24:11Z_
 
 ### Patch Changes
 
@@ -679,6 +849,8 @@
 
 ## 0.3.9
 
+_Released: 2026-03-27T15:37:24Z_
+
 ### Patch Changes
 
 - @sap-ux/fiori-generator-shared@0.13.93
@@ -687,6 +859,8 @@
 - @sap-ux/adp-tooling@0.18.106
 
 ## 0.3.8
+
+_Released: 2026-03-27T11:58:49Z_
 
 ### Patch Changes
 
@@ -701,6 +875,8 @@
 
 ## 0.3.7
 
+_Released: 2026-03-26T20:06:10Z_
+
 ### Patch Changes
 
 - Updated dependencies [b66e827]
@@ -714,6 +890,8 @@
 
 ## 0.3.6
 
+_Released: 2026-03-26T16:48:32Z_
+
 ### Patch Changes
 
 - Updated dependencies [96a689b]
@@ -721,12 +899,16 @@
 
 ## 0.3.5
 
+_Released: 2026-03-26T15:15:10Z_
+
 ### Patch Changes
 
 - Updated dependencies [3dcd3f7]
     - @sap-ux/adp-tooling@0.18.102
 
 ## 0.3.4
+
+_Released: 2026-03-26T12:07:04Z_
 
 ### Patch Changes
 
@@ -754,6 +936,8 @@
 
 ## 0.3.3
 
+_Released: 2026-03-25T12:56:41Z_
+
 ### Patch Changes
 
 - Updated dependencies [f384ace]
@@ -767,6 +951,8 @@
 
 ## 0.3.2
 
+_Released: 2026-03-23T18:25:40Z_
+
 ### Patch Changes
 
 - Updated dependencies [c0e05ab]
@@ -778,6 +964,8 @@
     - @sap-ux/abap-deploy-config-writer@0.3.1
 
 ## 0.3.1
+
+_Released: 2026-03-23T10:51:10Z_
 
 ### Patch Changes