+++ packages/ui-prompting/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui-prompting
 
+## 1.0.5
+
+### Patch Changes
+
+#### Dependency Updates
+
+- Upgrade patch-level dependencies [[aed328d](https://github.com/SAP/open-ux-tools/commit/aed328da8a5c93e226c58e4d7dc14c7c82756259)]
+
+#### Workspace Updates
+
+- @sap-ux/ui-components 3.0.3 → 3.0.4
+
 ## 1.0.4
 
 ### Patch Changes
@@ -10,12 +22,16 @@
 
 ## 1.0.3
 
+_Released: 2026-06-10T07:55:58Z_
+
 ### Patch Changes
 
 - Updated dependencies [a1235d6]
     - @sap-ux/ui-components@3.0.3
 
 ## 1.0.2
+
+_Released: 2026-06-03T14:58:37Z_
 
 ### Patch Changes
 
@@ -25,12 +41,16 @@
 
 ## 1.0.1
 
+_Released: 2026-06-03T10:37:52Z_
+
 ### Patch Changes
 
 - Updated dependencies [21cadc3]
     - @sap-ux/ui-components@3.0.1
 
 ## 1.0.0
+
+_Released: 2026-05-30T20:54:07Z_
 
 ### Major Changes
 
@@ -70,6 +90,8 @@
 
 ## 0.7.5
 
+_Released: 2026-05-29T09:10:40Z_
+
 ### Patch Changes
 
 - Updated dependencies [b0a922a]
@@ -77,12 +99,16 @@
 
 ## 0.7.4
 
+_Released: 2026-05-27T08:56:54Z_
+
 ### Patch Changes
 
 - Updated dependencies [bc315f0]
     - @sap-ux/ui-components@2.2.4
 
 ## 0.7.3
+
+_Released: 2026-05-26T16:40:21Z_
 
 ### Patch Changes
 
@@ -92,6 +118,8 @@
 
 ## 0.7.2
 
+_Released: 2026-05-22T06:20:19Z_
+
 ### Patch Changes
 
 - Updated dependencies [882d45b]
@@ -99,12 +127,16 @@
 
 ## 0.7.1
 
+_Released: 2026-05-15T10:52:26Z_
+
 ### Patch Changes
 
 - Updated dependencies [742d896]
     - @sap-ux/ui-components@2.2.1
 
 ## 0.7.0
+
+_Released: 2026-05-15T08:12:20Z_
 
 ### Minor Changes
 
@@ -117,6 +149,8 @@
 
 ## 0.6.29
 
+_Released: 2026-05-14T11:45:51Z_
+
 ### Patch Changes
 
 - 50a8ba5: chore: fresh release after workflow updates
@@ -125,12 +159,16 @@
 
 ## 0.6.28
 
+_Released: 2026-05-11T17:25:40Z_
+
 ### Patch Changes
 
 - Updated dependencies [189ad98]
     - @sap-ux/ui-components@2.1.15
 
 ## 0.6.27
+
+_Released: 2026-05-06T23:02:00Z_
 
 ### Patch Changes
 
@@ -139,12 +177,16 @@
 
 ## 0.6.26
 
+_Released: 2026-05-06T14:01:23Z_
+
 ### Patch Changes
 
 - Updated dependencies [7f8a712]
     - @sap-ux/ui-components@2.1.13
 
 ## 0.6.25
+
+_Released: 2026-04-30T14:23:24Z_
 
 ### Patch Changes
 
@@ -153,12 +195,16 @@
 
 ## 0.6.24
 
+_Released: 2026-04-22T14:26:13Z_
+
 ### Patch Changes
 
 - Updated dependencies [f77f003]
     - @sap-ux/ui-components@2.1.11
 
 ## 0.6.23
+
+_Released: 2026-04-14T21:37:37Z_
 
 ### Patch Changes
 
@@ -167,12 +213,16 @@
 
 ## 0.6.22
 
+_Released: 2026-04-01T10:25:12Z_
+
 ### Patch Changes
 
 - Updated dependencies [9357457]
     - @sap-ux/ui-components@2.1.9
 
 ## 0.6.21
+
+_Released: 2026-03-31T17:04:36Z_
 
 ### Patch Changes
 
@@ -181,6 +231,8 @@
 
 ## 0.6.20
 
+_Released: 2026-03-31T10:49:34Z_
+
 ### Patch Changes
 
 - Updated dependencies [f101671]
@@ -188,12 +240,16 @@
 
 ## 0.6.19
 
+_Released: 2026-03-31T08:19:05Z_
+
 ### Patch Changes
 
 - Updated dependencies [6b622e0]
     - @sap-ux/ui-components@2.1.6
 
 ## 0.6.18
+
+_Released: 2026-03-30T22:24:11Z_
 
 ### Patch Changes
 
@@ -208,12 +264,16 @@
 
 ## 0.6.17
 
+_Released: 2026-03-26T12:07:04Z_
+
 ### Patch Changes
 
 - Updated dependencies [a41533f]
     - @sap-ux/ui-components@2.1.4
 
 ## 0.6.16
+
+_Released: 2026-03-24T09:40:46Z_
 
 ### Patch Changes